requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap4==2024.1
cssselect2==0.7.0
Django==5.1.1
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx==0.2.4
fonttools==4.53.1
html5lib==1.1
lxml==5.3.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
webencodings==0.5.1
zopfli==0.2.3
